feat(edges): selectable with keys
